edit

+++ app/Http/Controllers/StudentController.php
@@ -52,27 +52,62 @@ class StudentController extends Controller
     public function update(Request $request , $id)
     {
         $student = Student::find($id);
-        $student -> first_Name = $request->input('first_Name');
-        $student -> last_Name = $request->input('last_Name');
-        $student -> email = $request->input('email');
-        $student -> industry = $request->input('industry');
-        if ($request->hasFile('profile_photo')) {
-            $destination = 'images/students/'.$student->profile_photo;
-            if (File::exists($destination)) {
-                File::delete($destination);
-            }
-            $file = $request->file('profile_photo');
-            $extention = $file->getClientOriginalExtension();
-            $filename = time().'.'.$extention;
-            $file->move('images/students/' , $filename);
-            $student->profile_photo = $filename;
-        }
-        $student -> update();
 
-        return response()->json([
-            'status' => 200,
-            'message' => 'Student Updated successfully',
+        if ($student) {
+
+            $student -> first_Name = $request->input('first_Name');
+            $student -> last_Name = $request->input('last_Name');
+            $student -> email = $request->input('email');
+            $student -> industry = $request->input('industry');
+            if ($request->hasFile('profile_photo')) {
+                $destination = 'images/students/'.$student->profile_photo;
+                if (File::exists($destination)) {
+                    File::delete($destination);
+                }
+                $file = $request->file('profile_photo');
+                $extention = $file->getClientOriginalExtension();
+                $filename = time().'.'.$extention;
+                $file->move('images/students/' , $filename);
+                $student->profile_photo = $filename;
+            }
+            
+            $student ->update();
+
+            return response()->json([
+                'status' => 200,
+                'message' => 'Student Added successfully',
+            ]);
+
+        } else {
+            return response()->json([
+            'status' => 404,
+            'message' => 'Student not found',
         ]);
+        }
+
+
+        // $student = Student::find($id);
+        // $student -> first_Name = $request->input('first_Name');
+        // $student -> last_Name = $request->input('last_Name');
+        // $student -> email = $request->input('email');
+        // $student -> industry = $request->input('industry');
+        // if ($request->hasFile('profile_photo')) {
+        //     $destination = 'images/students/'.$student->profile_photo;
+        //     if (File::exists($destination)) {
+        //         File::delete($destination);
+        //     }
+        //     $file = $request->file('profile_photo');
+        //     $extention = $file->getClientOriginalExtension();
+        //     $filename = time().'.'.$extention;
+        //     $file->move('images/students/' , $filename);
+        //     $student->profile_photo = $filename;
+        // }
+        // $student -> update();
+
+        // return response()->json([
+        //     'status' => 200,
+        //     'message' => 'Student Updated successfully',
+        // ]);
     }
 
     public function destroy($id)
